Multiple Updates
Added Classes:
We finished the TriangleMesh class and the Triangle Class. They appear
to be working but we haven't checked all of the functionality.

Changed Interface:
We removed the getNormalAt() function from
AbstractSceneObject/SceneObject. This may break the current
simpleRayTracer code. The normal class of the intersection object
should contain the normal now and should be used. Have not debugged
this field yet.